event_notifier: enable it to use pipes
This takes the eventfd emulation code from the main loop.
When the EventNotifier is used for the main loop too, we
need this compatibility code.

event_notifier_init_fd and event_notifier_get_fd are used
for low-level code talking to KVM, so make them dependent
on the "real" eventfd.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>