separate thread for VM migration
This patch creates a separate thread for the guest migration on the
source side.

BufferedFile is changed to push as much data as possible (within the
rate limit) on a blocking socket, and pause for at most 100 ms when
the limit is exceeded.  Buffering is necessary because it allows to do
I/O outside locks.  The put_ready callback is called without locks,
and takes the iothread or ramlist locks only when necessary.

qemu_fclose of the BufferedFile must be invoked from the iothread,
because it blocks synchronously until the migration thread is done with
the current iteration and terminates its execution.  I use a bottom
half to do this from the migration thread when migration is finished
(migrate_fd_close is called from the migration thread; migrate_fd_release
from the I/O thread, typically from the monitor).

Because of this, migrate_cancel requests from the iothread are also
handled synchronously.

Signed-off-by: Umesh Deshpande <udeshpan@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>