main-loop: main_loop_wait returns 1 if bottom halves were run
This will be needed by qemu_aio_flush/qemu_aio_wait.  It is fine because
if ret was zero and yet bottom halves were triggered (e.g. by timers)
there has been a call to qemu_notify_event, and it is worthwhile to
clean up the eventfd without relinquishing (in the TCG case) the global
mutex.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>