aio: do not make rfds/wfds conditional on io_flush result
Instead, assume that io_read and io_write will make progress when they
trigger.  For io_read, if there is data ready it shall either read it,
or disable itself.  For io_write handler, it shall not be set unless it
will actually have something to write.  This is true for all callers.

With this in place, abstract the io_flush walk into a new function, and
use it from both qemu_aio_wait and qemu_aio_flush.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>