lockless access to aiocb->ret and aiocb->queued
aiocb->ret and aiocb->queued can be accessed lockless if we introduce
a "special" value for aiocb->ret that is used while aiocb->queued = 1.

Thus, the update sequence is

    -ENOTSTARTED    queued = 1
    -EINPROGRESS    queued = 1
    -EINPROGRESS    queued = 0
    <return code>   queued = 0

Upon cancellation, what happens is:

    -ENOTSTARTED    queued = 1
    -ECANCELED      queued = 1 (by paio_cancel)
    -EINPROGRESS    queued = 1 (but ECANCELED is saved via atomic_xchg)
    -EINPROGRESS    queued = 0
    -ECANCELED      queued = 0

Thanks to ENOTSTARTED, paio_cancel can signal cancellation only on an
inactive aiocb, atomically and without double-word cmpxchg.  With two
concurrent cancellations, the second one will fail but will exit
either immediately (if it finds -ECANCELED) or very soon (if it
finds -EINPROGRESS).

Cleaning up canceled aiocbs is done always in posix_aio_process_queue,
it is not a fast path anyway.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>