mirror: introduce mirror job
This patch adds the implementation of a new job that mirrors a disk to
a new image while letting the guest continue using the old image.
The target is treated as a "black box" and data is copied from the
source to the target in the background.

The job is possibly never-ending, but it is logically structured into
two phases: 1) copy all data as fast as possible until the target
first gets in sync with the source; 2) keep target in sync and
ensure that reopening to the target gets a correct (full) copy
of the source data.

The second phase is indicated by the progress in "info block-jobs"
reporting the current offset to be equal to the length of the file.
When the job is cancelled in the second phase, QEMU will run the
job until the source is clean and quiescent, then it will report
successful completion of the job.  (Note that it could already happen
that management lost the race against QEMU and got a completion
event instead of cancellation).

One borderline case that is not handled here (and fails silently) is
when the target (and only the target) has clusters that are bigger than
1MB, the dirty bitmap granularity.  In this case, opening the file with
BDRV_O_NO_BACKING is wrong and can leave zeros replacing parts of the
source file.  This is fixed later, and an interim solution would be
relatively complicated.

As well as fixing this, the next patches will introduce many refinements
to the basic idea introduced here, including an asynchronous event to
signal the phase switch, a new command to complete the job and switch over
to the target disk, some tunable knobs and many performance optimizations.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>