nbd: do not leak nbd_trip coroutines when a connection is torn down
Because nbd_client_close removes the I/O handlers for the client
socket, there is no way that any suspended coroutines are restarted.
This will be a problem with the QEMU embedded NBD server, because
we will have a QMP command to forcibly close all connections with
the clients.

Instead, we can exploit the reference counting of NBDClients; shutdown
the client socket, which will make it readable and writeable.  The
coroutines then will fail and exit cleanly.  The last refcount finally
triggers the closure of the client.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>